Fix broken imports, version bumps
This commit includes some maintenance tasks that were needed in order
to get the DAGs in this repo to work again in the local Astronomer
setup. Specifically:
1. Restrict the versions of packages in use to avoid this happening
again (part of the problem was that `astronomer-cosmos` was unpinned,
leading it to grab the latest version, which then caused compatibility
issues with the existing code).
2. Bump versions where appropriate
3. Rework DAGs to use the new `Config` classes (e.g. `RenderConfig`,
`ProfileConfig`, `ExecutionConfig`.
4. Remove the virtualenv in the Dockerfile, we can simplify by using
the Python environment provided in the Docker container.
5. Fix a bug in the dbt code (see [this commit](https://github.com/astronomer/astronomer-cosmos/pull/516/commits/deb99649ddf7cd656e3b4641f050012dc54f8530#diff-04ddeb3f5adbe4a0b02dc99db61196647086840e023b90ceecdbb4e4ce72988cR1))
that caused failing dbt jobs when the materialization was set to `table`.